requirements.txt
absl-py==2.0.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
astunparse==1.6.3
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
cssselect2==0.7.0
Django==4.2.7
django-autoslug==1.9.9
flatbuffers==23.5.26
fonttools==4.44.0
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
html5lib==1.1
idna==3.4
joblib==1.3.2
keras==2.14.0
libclang==16.0.6
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==23.2
Pillow==10.1.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.8.0
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pypdf==3.17.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.0.7
weasyprint==60.1
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.6.0
wrapt==1.14.1
xhtml2pdf==0.2.11
zopfli==0.2.3
